# MyCommute.Shared/Enums/CommuteType.cs
namespace MyCommute.Shared.Enums
{
    public enum CommuteType
    {
        Undefined,
        Bike,
        Car,
        CarpoolDriver,
        CarpoolPassenger,
        PublicTransport,
        Train,
        Walking,
    }
}